backend/pyproject.toml
[project]
name = "webapp-telegram-fastapi"
version = "0.1.0"
description = "Web application skeleton for building a Telegram bot using FastAPI and Python Telegram Bot"
readme = "README.md"
requires-python = "<4.0,>=3.9"
dependencies = [
    "beanie>=2.0.0",
    "fastapi>=0.118.0",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pymongo>=4.15.2",
    "python-jose>=3.5.0",
    "python-telegram-bot>=22.5",
    "uvicorn>=0.37.0",
]
